backend/requirements.txt
fastapi==0.115.13
uvicorn[standard]==0.32.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
cryptography==43.0.3
boto3==1.35.50
reportlab==4.2.5
pydantic==2.9.2
pydantic-settings==2.6.1
sqlalchemy==2.0.36
alembic==1.14.0
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.27.2
requests==2.32.3
respx==0.22.0
authlib==1.3.1
